test: separate DependencyAssembly into TypeDependencyAssembly and MethodDependencyAssembly
