+++ ArchUnitNETTests/ArchUnitNETTests.csproj
@@ -17,7 +17,8 @@
     <ProjectReference Include="..\TestAssemblies\InterfaceAssembly\InterfaceAssembly.csproj" />
     <ProjectReference Include="..\TestAssembly\TestAssembly.csproj" />
     <ProjectReference Include="..\TestAssemblies\AttributeAssembly\AttributeAssembly.csproj" />
-    <ProjectReference Include="..\TestAssemblies\DependencyAssembly\DependencyAssembly.csproj" />
+    <ProjectReference Include="..\TestAssemblies\TypeDependencyAssembly\TypeDependencyAssembly.csproj" />
+    <ProjectReference Include="..\TestAssemblies\MethodDependencyAssembly\MethodDependencyAssembly.csproj" />
     <ProjectReference
       Include="..\TestAssemblies\FilteredDirectoryLoaderTestAssembly\FilteredDirectoryLoaderTestAssembly.csproj"
       ReferenceOutputAssembly="false"
